normed and merged
