+++ Makefile
@@ -11,6 +11,7 @@ LIBFT_PATH = libft/
 MINISHELL_INCS= 						\
 	src/defines.h						\
 	src/syntax/check_syntax.h			\
+	src/syntax/redirection_syntax.h		\
 	src/parser/parser.h					\
 	src/parser/command_table.h			\
 	src/parser/parse_redirection.h		\
@@ -31,6 +32,7 @@ MINISHELL_INCS= 						\
 
 MINISHELL_SRC= 							\
 	src/syntax/check_syntax.c			\
+	src/syntax/redirection_syntax.c		\
 	src/parser/parser.c					\
 	src/output/prompt.c					\
 	src/output/write_to_std.c			\
