+++ src/output/write_to_std.c
