+++ src/output/write_to_std.h
@@ -1,5 +1,5 @@
 #ifndef WRITE_TO_STD_H
-#define WRITE_TO_STD_H
+# define WRITE_TO_STD_H
 
 void	write_to_stdout(const char *string_to_write);
 void	write_to_stderr(const char *string_to_write);
